backend/requirements.txt
fastapi>=0.109.1
starlette>=0.36.2
uvicorn==0.25.0
gunicorn==22.0.0
requests==2.31.0
streamlit==1.30.0
sqlalchemy==2.0.24
numpy==1.24.3
bcrypt==4.1.2
pyyaml==6.0
alembic==1.13.1
pymysql==1.1.0
cryptography==42.0.4
beautifulsoup4==4.12.3
lxml==5.1.0
selenium==4.18.1
webdriver-manager==4.0.1
pytest-mock==3.12.0
httpx==0.23.1
python-jose==3.3.0
itsdangerous==2.2.0
redis==5.0.4
requests-mock==1.12.1
fakeredis==2.23.2
tenacity==8.3.0
